requirements.txt
google-cloud-storage
unittest2